+++ src/components/views/Home/CarouselComments/CarouselComments.jsx
@@ -1,0 +1,8 @@
+// import style from "../CarouselComments/CarouselComments.module.css"
+
+
+const CarouselComments = () => {
+
+};
+
+export default CarouselComments;
