+++ src/components/views/Home/CarouselSale/CarouselSale.module.css
@@ -1,0 +1,59 @@
+.carousel {
+    width: 78%;
+    margin: 5px auto;
+    margin-bottom: 15px;
+  }
+  
+  .card {
+    box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.5);
+    max-width: 300px;
+    margin: auto 1rem;
+    text-align: center;
+    /* font-family: arial; */
+  
+    /* Nuevos estilos para asegurar el mismo tamaño de las tarjetas */
+    display: flex;
+    flex-direction: column;
+    justify-content: space-between;
+    height: 100%;
+  }
+  
+  .productImage {
+    padding-top: 5px;
+    height: 300px;
+    width: 95%;
+    padding: 40px;
+  }
+  
+  .priceBefore {
+    color: red;
+    font-size: 18px;
+  }
+  
+  .priceNow {
+    margin-top: -15px;
+    color: green;
+    font-size: 20px;
+  }
+  
+  .card button {
+    border: none;
+    outline: 0;
+    padding: 12px;
+    color: white;
+    background-color: #FF9E5C;
+    text-align: center;
+    cursor: pointer;
+    width: 100%;
+    font-size: 18px;
+    border-color: #FF9E5C;
+    border-style: solid;
+  }
+  
+  .card button:hover {
+    background-color: white;
+    color:#FF9E5C;
+    border-color: #FF9E5C;
+    border-style: solid;
+  }
+
