+++ src/components/views/Home/Home.jsx
@@ -1,6 +1,4 @@
 import React from "react";
-import Carousel from 'react-multi-carousel';
-import 'react-multi-carousel/lib/styles.css';
 import { useEffect } from "react";
 import { useDispatch, useSelector } from "react-redux";
 import { Link } from "react-router-dom";
@@ -8,19 +6,17 @@ import { useAuth0 } from "@auth0/auth0-react";
 import {getProducts, getProductsSales, getProductsRating, getById, getProductsOnHome, showProductsSalesOnShop, showProductsRatingOnShop } from "../../../redux/productActions";
 import { postUserToBack } from "../../../redux/UserActions";
 import style from "../Home/Home.module.css"
+import CarouselSale from "./CarouselSale/CarouselSale";
+import CarouselRating from "./CarouselRating/CarouselRating";
 
 
 
 const Home = () => {
   const dispatch = useDispatch()
-  const { productsOnSale } = useSelector(state => state.products)
-  const { productsOnRating } = useSelector(state => state.products)
   const { postedUser } = useSelector(state => state.user);
 
   const { isAuthenticated, user } = useAuth0();
 
-  const productsSales = productsOnSale.slice(0, 12)
-  const productsRating = productsOnRating.slice(0, 12)
 
   useEffect(() => {
     dispatch(getProductsSales())
@@ -36,26 +32,6 @@ const Home = () => {
   }, [dispatch, isAuthenticated, postedUser])
 
 
-  const responsive = {
-    superLargeDesktop: {
-      // the naming can be any, depends on you.
-      breakpoint: { max: 4000, min: 3000 },
-      items: 5
-    },
-    desktop: {
-      breakpoint: { max: 3000, min: 1024 },
-      items: 4
-    },
-    tablet: {
-      breakpoint: { max: 1024, min: 464 },
-      items: 2
-    },
-    mobile: {
-      breakpoint: { max: 464, min: 0 },
-      items: 1
-    }
-  };
-
   const clickDispatch = (id) => {
     dispatch(getById(id))
   }
@@ -70,51 +46,17 @@ const Home = () => {
 
   return (
     <div className={style.container}>
-      {/*----------------- PRODUCTOS EN OFERTA -------------------*/}
-
+      {/*------------------------------------------- CARRUSEL PRODUCTOS EN OFERTA -----------------------------------------*/}
       <h2 className={style.h2}>Productos en oferta <Link to={"/shop"} className={style.link}><span onClick={() => dispatchProductsSalesToShop()} className={style.span}>Ver todos</span></Link></h2>
 
-      <Carousel responsive={responsive} className={style.carousel} infinite={true} >
-        {
-          productsSales && productsSales.map((product, index) => {
-            return (
-              <div key={index} className={style.card} >
-                <img className={style.productImage} src={product.image} alt={product.name} />
-                <h5>{product.name}</h5>
-                <p className={style.priceBefore}>Antes: ${product.price}</p>
-                <p className={style.priceNow}>Ahora: ${product.salePrice}</p>
-                <Link to={`/shop/${product._id}`}>
-                  <button onClick={() => clickDispatch(product._id)} >Ver producto</button>
-                </Link>
-              </div>
-            )
-          })
-        }
-      </Carousel>
+      <CarouselSale clickDispatch={clickDispatch} />
 
-      {/*---------------- PRODUCTOS POR RATING ----------------- */}
-
+      {/*------------------------------------------- CARRUSEL PRODUCTOS POR RATING ---------------------------------------- */}
       <h2 className={style.h2}>Productos mas vendidos <Link to={"/shop"} className={style.link}><span onClick={() => dispatchProductsRatingToShop()} className={style.span}>Ver todos</span></Link></h2>
-      <Carousel responsive={responsive} className={style.carousel} infinite={true} >
-        {
-          productsRating && productsRating.map((product, index) => {
-            return (
-              <div key={index} className={style.card} >
-                <img className={style.productImage} src={product.image} alt={product.name} />
-                <h5>{product.name}</h5>
-                <p className={style.rating}>Rating: {product.rating}</p>
-                <p className={style.priceNow}>Precio: ${product.price}</p>
-                <Link to={`/shop/${product._id}`}>
-                  <button onClick={() => clickDispatch(product._id)} >Ver producto</button>
-                </Link>
-              </div>
-            )
-          })
-        }
-      </Carousel>
 
+      <CarouselRating clickDispatch={clickDispatch} />  
 
-
+      {/*------------------------------------------- FALTA CARRUSEL DE COMENTARIOS ---------------------------------------- */}   
 
     </div>
   )
